+++ packages/harpstrata/src/harpstrata/Covariant/deduceRootPitchId/index.ts
@@ -1,0 +1,16 @@
+import { RootPitchControlVars } from '../types'
+import { getPozitionRootOffset } from '../../Pozition'
+import type { PitchIds } from '../../Pitch'
+import { getAscendingPitchIds } from '../../Pitch'
+
+
+export const deduceRootPitchId = (props: RootPitchControlVars): PitchIds => {
+  const { harpKeyId, pozitionId } = props
+  const rootOffset = getPozitionRootOffset(pozitionId)
+
+  const ascendingPitchIds = getAscendingPitchIds(harpKeyId)
+
+  const { [rootOffset]: rootPitchId } = ascendingPitchIds
+
+  return rootPitchId
+}