+++ packages/harpstrata/src/harpstrata/Covariant/index.ts
@@ -1,0 +1,2 @@
+export { getCovariants } from './getCovariants'
+export type { CovariantControlVars, HarpKeyControlVars, RootPitchControlVars, PozitionControlVars } from './types'