+++ packages/harpstrata/src/harpstrata/HarpStrata/getHarpStrata/index.ts
@@ -1,27 +1,26 @@
 import type { HarpStrataProps, HarpStrata } from '../types'
-import { getRootPitch } from '../getRootPitch'
-import { getPozition } from '../../Pozition'
+import { getPozitionRootOffset } from '../../Pozition'
 import { getPitchMatrix, getPitch } from '../../Pitch'
 import { getIsActiveComplex } from '../../IsActive'
 import { getDegreeMatrix } from '../../Degree'
+import { getCovariants } from '../../Covariant'
 import { getApparatus } from '../../Apparatus'
 
 export const getHarpStrata = (props: HarpStrataProps): HarpStrata => {
   const { apparatusId, pozitionId, harpKeyId, activeIds } = props
 
   const apparatus = getApparatus(apparatusId)
-  const pozition = getPozition(pozitionId)
+  const pozitionRoot = getPozitionRootOffset(pozitionId)
   const pitch = getPitch(harpKeyId)
 
   const { halfstepIndexMatrix } = apparatus
-  const { root: pozitionRoot } = pozition
   const { id: pitchId } = pitch
 
   const degreeMatrix = getDegreeMatrix(halfstepIndexMatrix, pozitionRoot)
   const pitchMatrix = getPitchMatrix(halfstepIndexMatrix, pitchId)
   const isActiveComplex = getIsActiveComplex({degreeMatrix, pitchMatrix, activeIds: activeIds })
 
-  const rootPitchId = getRootPitch(degreeMatrix, pitchMatrix)
+  const { rootPitchId } = getCovariants({pozitionId, harpKeyId})
 
   return {
     apparatus,