+++ packages/harpstrata/src/harpstrata/HarpStrata/index.ts
@@ -1,13 +1,13 @@
-import { getActivePozitionIds } from '../Pozition'
-import { getActivePitchIds } from '../Pitch'
+import { getAscendingPozitionIds } from '../Pozition'
+import { getAscendingPitchIds } from '../Pitch'
 import { getActiveApparatusIds } from '../Apparatus'
 
 
 export const getApparatusIds = getActiveApparatusIds
 
-export const getPozitionIds = getActivePozitionIds
+export const getPozitionIds = getAscendingPozitionIds
 
-export const getPitchIds = getActivePitchIds
+export const getPitchIds = getAscendingPitchIds
 
 export { getHarpStrata } from './getHarpStrata'