+++ packages/harpstrata/src/harpstrata/Pitch/getPitchMatrix/index.ts
@@ -1,0 +1,15 @@
+import type { PitchIds, PitchMatrix } from '../types'
+import { getAscendingPitchIds, getPitch } from '../pitchMap'
+import type { HalfstepIndexMatrix } from '../../Apparatus'
+
+export const getPitchMatrix = (halfstepIndexMatrix: HalfstepIndexMatrix, keyPitchId: PitchIds): PitchMatrix => {
+
+  const ascendingPitchIdsInKey = getAscendingPitchIds(keyPitchId)
+
+  return halfstepIndexMatrix.map((halfstepIndexRow) => {
+    return halfstepIndexRow.map((halfstepIndex) => {
+      if (halfstepIndex === undefined) return undefined
+      return getPitch(ascendingPitchIdsInKey[halfstepIndex % 12])
+    })
+  })
+}