+++ packages/harpstrata/src/harpstrata/Pozition/types/index.ts
@@ -1,5 +1,3 @@
-import type { HalfstepIndex } from '../../Apparatus'
-
 export enum PozitionIds {
   First = 'FIRST',
   Second = 'SECOND',
@@ -17,5 +15,4 @@ export enum PozitionIds {
 
 export type Pozition = {
   readonly id: PozitionIds;
-  readonly root: HalfstepIndex;
 }