+++ packages/harpstrata/src/harpstrata/index.ts
@@ -17,6 +17,9 @@ export type { Pozition } from './Pozition'
 export { PitchIds } from './Pitch'
 export type { Pitch } from './Pitch'
 
+export { getCovariants } from './Covariant'
+export type { CovariantControlVars, HarpKeyControlVars, RootPitchControlVars, PozitionControlVars } from './Covariant'
+
 export { IsActiveIds } from './IsActive'
 export type {
   ActiveDegreeIds, ActivePitchIds, ActiveIds,