+++ packages/harpstrata/CHANGELOG.md
@@ -12,7 +12,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Security: to invite users to upgrade in case of vulnerabilities.
 
 
-## [Unreleased](https://github.com/js-jslog/harpstrata/compare/v5.0.1...HEAD) - yyyy-mm-dd
+## [Unreleased](https://github.com/js-jslog/harpstrata/compare/v5.1.0...HEAD) - yyyy-mm-dd
+## [v5.1.0](https://github.com/js-jslog/harpstrata/releases/tag/v5.1.0)
+### Added
+- MINOR: Add covariant deduction functions
+- MINOR: Add ability to specify origin value on pitch & pozition id list getters
+- PATCH: Pitch, Pozition and Degree id ordered list generation (for internal use)
+
+### Removed
+- PATCH: ORDERED_XXX constants. Replaced by the ordered list generation described above (internal use only)
+
 ## [v5.0.0](https://github.com/js-jslog/harpstrata/releases/tag/v5.0.0) - 2020-05-17
 ### Added
 - MAJOR: Add all the properties which the calling app would need to know how the harpstrata was generated
@@ -62,7 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Library created
 
 ## Github release list
-- [unreleased](https://github.com/js-jslog/harpstrata/compare/v5.0.0...HEAD)
+- [unreleased](https://github.com/js-jslog/harpstrata/compare/v5.1.0...HEAD)
+- [v5.1.0](https://github.com/js-jslog/harpstrata/releases/tag/v5.1.0)
 - [v5.0.0](https://github.com/js-jslog/harpstrata/releases/tag/v5.0.0)
 - [v4.0.1](https://github.com/js-jslog/harpstrata/releases/tag/v4.0.1)
 - [v4.0.0](https://github.com/js-jslog/harpstrata/releases/tag/v4.0.0)